refactor: remove duplicate code when cluster merging
